fix: bump rand to 0.8.6/0.9.3 (GHSA-cq8v-f236-94qc) + Linux clippy casts
Two fixes bundled because they were both caught during the same
CI-red investigation:

1) Dependabot alerts #1 (rand 0.9.x path via proptest dev-dep) and
   #2 (rand 0.8.x direct dep) — both GHSA-cq8v-f236-94qc, "rand is
   unsound with a custom logger using rand::rng()". Low severity,
   trigger requires all of: log + thread_rng features, custom
   logger, that logger calling rand::rng() and hitting the 64 kB
   reseed boundary, trace-level or warn+getrandom-failure. GVM
   uses tracing (not log), no code path matches. Still bumping
   the lock so the alerts close:
     - rand 0.8.5 → 0.8.6 (Cargo.toml semver "0.8" allows)
     - rand 0.9.2 → 0.9.3 (transitive via proptest, dev-only)
   Cargo.toml unchanged; only Cargo.lock is repinned.

2) Linux clippy has been red on master since 60c241c
   (the commit that added --all-targets). Root cause is the same
   pattern as the bench-fixture drift: cargo clippy on Windows
   skips #[cfg(target_os = "linux")] blocks entirely, so
   Linux-only code in crates/gvm-sandbox/src/syscall_names.rs
   was never linted locally. The failing lints are five
   clippy::unnecessary_cast in the test module:

     let n = libc::SYS_mount as i64;   // SYS_* is already i64 on Linux
     let n = libc::SYS_ptrace as i64;
     let n = libc::SYS_bpf as i64;
     let n = libc::SYS_read as i64;
     let n = libc::SYS_openat as i64;

   Removing the casts is the fix — libc::SYS_* is `c_long` (i64
   on 64-bit Linux, i32 on 32-bit) and our proxy is 64-bit only.
   Behavioural no-op; the cast was redundant.

Bigger-picture takeaway (not fixed here): `cargo clippy` on any
single OS can only see the cfg branches that compile on that OS.
Fully closing the loop would need a CI matrix (Linux+macOS+Windows)
on the clippy job. Punting on that until the next platform-only
lint bites — the pattern is now documented in ci.yml.

Unrelated: the Nightly Stress (EC2) failure on 2026-06-28 is a
missing-secrets issue (EC2_HOST / EC2_SSH_KEY not configured), not
a code bug — the workflow's sanity-check step is designed to bail
early when secrets are absent. No fix needed here; the operator
sets those secrets when they want the scheduled stress test to run.

+++ crates/gvm-sandbox/src/syscall_names.rs
@@ -340,19 +340,19 @@ mod tests {
     #[test]
     fn lookup_blocked_mount() {
         // mount(2) is the canonical example used in our seccomp filter.
-        let n = libc::SYS_mount as i64;
+        let n = libc::SYS_mount;
         assert_eq!(name_for(n), Some("mount"));
     }
 
     #[test]
     fn lookup_blocked_ptrace() {
-        let n = libc::SYS_ptrace as i64;
+        let n = libc::SYS_ptrace;
         assert_eq!(name_for(n), Some("ptrace"));
     }
 
     #[test]
     fn lookup_blocked_bpf() {
-        let n = libc::SYS_bpf as i64;
+        let n = libc::SYS_bpf;
         assert_eq!(name_for(n), Some("bpf"));
     }
 
@@ -360,7 +360,7 @@ mod tests {
     fn lookup_allowed_read() {
         // Allowed syscalls are also in the table so error messages stay
         // useful when an agent dies on an unexpected syscall.
-        let n = libc::SYS_read as i64;
+        let n = libc::SYS_read;
         assert_eq!(name_for(n), Some("read"));
     }
 
@@ -373,7 +373,7 @@ mod tests {
     #[test]
     fn name_strips_sys_prefix() {
         // The macro strips "SYS_" so output is human-friendly.
-        let n = libc::SYS_openat as i64;
+        let n = libc::SYS_openat;
         assert_eq!(name_for(n), Some("openat"));
     }
 }
